Fixed GifDrawable's toString() method
The method was passing the wrong arguments into the string formatter
making it appear there was something wrong with the GIF's header data.

This is part of a larger effort to track down a segfault that only
occurs when the GifDrawable is constructed via an InputStream rather
than a FileDescriptor.