Gracefully skip bad image blocks when parsing
Changing the order of operations within DGifGetImageDesc so that we can
discard bad image blocks and continue gracefully.

Prior to this change, the bad image was still read into SavedImages,
leading to a inconsistency in image size metadata in SavedImage vs
GifFileType. This causes potential read/write out of bounds that may
cause app crashes.