Fix buffer size allocation with malformed gifs
Previously, if either widthOverflow or heightOverflow were negative
and the other was positive, this would result in adding a negative
number to gifFilePtr's dimensions and therefore make us allocate
too small of a buffer resulting in an OOB heap overflow.

This patch separates the negative number check to be per dimension
and resolves the issue.